Little Buggy Rotation Adjustment
Okay it is not something that I'm really proud of but you can turn shapes and put them in the grids but there are small ISSUES about reseting shape types. If you reset shapes consecutively you will end up encountering not correct rotations and non-existing shapes on the white check shape screen but 2nd problem disappears when you put a shape to the girds. If you press rotate and reset in order and repeat it, there is no problem at all. I have no idea what is going on at this point. If you fix it I would really appreciate. If not, I will fix it later in this night.