Reset Shapes and Rotating relation bug fix
It is not completely finished yet, you can reset your shapes up to three times flawlessly. If you want to do it more than 3 times repeatedly it won't work so please do not :P also non-existing and not correct shapes still exist but if you reset your shapes and rotate them just after that it will be perfectly fine. However you can't do it in the codes it is not gonna work because... I also don't know if you figure it out let me know that.